Don't decorate the content package with chunk methods on every intialise
Every instance of a content package was being decorated with the
content chunk methods when it was initialised. This was problematic for
performance especially building the tree select as the dynamic methods
were never actually called.

Instead we build the chunks the first time a method is called on the
content package that isn't defined yet