Add a leading slash to resource links
When using the Resource content attribute type, only the id of the content
package to link to is stored.
The content package is then decorated with an *_path method which gets the
latest permalink for that content package.

This was returning urls without a leading slash:
    "this-is-an/example/url"
This causes problems as the url is appended to the existing url when linked.

This commit prepends a leading slash to the url to ensure the link can always be
directly linked without any manipulation.